ai-study-assistant/requirements.txt
streamlit==1.56.0
requests==2.33.1
python-dotenv==1.2.2
numpy==2.2.6
endee==0.1.25